refactor: controller와 관련한 컴포넌트를 "controller" 패키지로 이동
